requirements.txt
streamlit
google-api-python-client
youtube-transcript-api
pandas
requests
pyairtable
yt-dlp
openai-whisper
selenium
beautifulsoup4
firebase-admin
python-dotenv